pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "so2x-harness"
version = "0.6.2"
description = "Thin personal per-project harness"
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
]

[project.scripts]
so2x = "cli.main:main"

[tool.setuptools.package-dir]
"" = "scripts"

[tool.setuptools.packages.find]
where = ["scripts"]
include = ["cli*"]

[tool.setuptools]
py-modules = ["doctor", "runtime"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.coverage.run]
parallel = true
source = ["scripts"]
patch = ["subprocess"]

[tool.coverage.paths]
source = [
    "scripts",
    "*/so2x-harness/scripts",
]
